requirements.txt
bokeh
altair
streamlit
vega_datasets 
matplotlib.pyplot
